syntax fixes

+++ internal/discovery/hazelcast_cloud_discovery.go
@@ -118,7 +118,7 @@ func (hzC *HazelcastCloud) parseResponse(response *http.Response) (map[string]co
 		publicAddress := hzC.createAddress(addr.PublicAddr)
 		// TODO:: what if privateAddress is not okay ?
 		// TODO:: use addressProvider
-		privateAddress := proto.NewAddressWithParameters(addr.PrivAddr, publicAddress.Port())
+		privateAddress := proto.NewAddressWithParameters(addr.PrivAddr, int32(publicAddress.Port()))
 		privateToPublicAddrs[privateAddress.String()] = publicAddress
 	}
 
@@ -132,6 +132,6 @@ func CreateURLEndpoint(hazelcastProperties *property.HazelcastProperties, cloudT
 
 func (hzC *HazelcastCloud) createAddress(hostname string) core.Address {
 	ip, port := iputil.GetIPAndPort(hostname)
-	addr := proto.NewAddressWithParameters(ip, port)
+	addr := proto.NewAddressWithParameters(ip, int32(port))
 	return addr
 }
